Movement update
Write two scripts based off of the 2D assets, RifterControl and
RifterControlInterface. Added the character prefab to our own prefab
list because it now uses my scripts. BlackHoleMaker was not actually
changed, I added a single space, sorry. The vortex prefabs were changed
so that they are a part of the new "holes" layer upon creation, meaning
the control scripts don't recognize them as ground for jumping on.