Camera change
I didn't actually change any prefabs, or scenes, because that's a
headache that I don't want to deal with at this point. The only things
contained in this change are newly created things.

IMPORTANT NOTE: THE CAMERA SCRIPT REQUIRES THAT THE MAIN CAMERA IS
UNASSOCIATED WITH THE CHARACTER, YOU HAVE TO CHANGE THE CHARACTER TO
UNATTACH THE CAMERA.