Pretty big changes, be careful
I changed a bunch of things. The things that aren't mine  include the
black and white hole prefabs, the exit portal, and the "final scene".
However, I built on the second level a lot and I created some more
scripts to make things look flashy. It's not quite done though, the
Spontaneous White Hole script is not functional and I'll probably make
one more commit before I'm done for the night.